Changed directory organization, implemented model.py methods into app.py, clean up
used the following directory format :
dataset :
└── demo
      └── keylog_.....csv
      └──audio_...wav
....
└── <other development_dataset>

This simplifies the combination of demo and development dataset generation

Also, I set the saving of numpy spectrograms as optional. Instead, the function generate_spectrograms returns a list of ndarray directly, which is later used for training (simpler).

There are currently issues in the conversion of List[ndarray] -> List[torch.Tensor] but I'll fix this tomorrow